Add resource type 'MSS universal driver' to patch files and install-ICAT-from-scratch SQL.
Add the ora-patch-v2.1tov2.2.sql patch file.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@2114 ea137bd6-aa86-47b6-b90d-b101d9f19a90

+++ iRODS/server/icat/patches/ora-patch-v2.1tov2.2.sql
@@ -1,10 +1,10 @@
--- run these SQL statements, using the MySQL client mysql
+-- run these SQL statements, using the Oracle client sqlplus
 
 alter table R_RESC_MAIN add column resc_status varchar(32);
 
 create table R_USER_AUTH
 (
-   user_id             bigint not null,
+   user_id             integer not null,
    user_auth_name      varchar(1000),
    create_ts varchar(32)
 );
